+++ MaXrd/Changelog.md
@@ -1,5 +1,11 @@
 # _MaXrd_: Mathematica X-ray diffraction package – change log
 
+## Version 1.4.0
+- Renamed the options `"RandomDistortions"` and `"RandomRotations"` to `"Distortions"` and `"Rotations"`, respectively, in `EmbedStructure` and made them more general by enabling the choice between set values or ranges.
+- Added the option `"DistortionType"` which enables the function to interpret the given distortions either as ångströms in a `"Cartesian"` system or as fractions of the host unit cell in a `"Crystallographic"` setting.
+- Added the option `"RotationOrder"` which lets the user specify the order of axes to rotate.
+
+
 ## Version 1.3.0
 - Updated the example under _Scope_ in the `UnitCellTransformation` documentation page to make use of `CrystalPlot`.
 - Added `EmbedStructure` to the list in `AutoComplete.nb`.